Move Qt widgets to new module
This encapsulates code that needs to be updated by changes to the Matplotlib widgets, in particular to the ColorButton class now defined in the new formlayout package.
